Feat/deepseek non absorption (#941)
* deepseek MLA

* Feat/attention sink (#207)

* feat: add RPA v3 kernel with tpu-inference optimizations and attention sink

    - Port RPA v3 kernel with all 9 tpu-inference optimizations
    - Support attention sink, custom mask, xai_temperature, sliding window
    - Update ref function and tests to use v3 as baseline
    - Switch benchmark to use v3 kernel


* wrap kv cache update with shard map

* fix: attention_sink TP sharding and MHA+bf16 q_packing alignment (#220)

- Shard attention_sink by tensor axis for TP compatibility
- Move attention_sink preparation into prepare_inputs, pad per kv_head
  then pad q_heads_per_kv dimension to match kernel's padded layout
- Restructure l/m initialization to avoid redundant writes
- Add attention_sink tests: MHA, GQA(4q/1kv), single-head(1q/1kv)
- Update test_flashattention_dp.py to use v3 ref and support attention_sink


* feat: change kv cache shape to match rpav3 (#228)

* tune rpav3 (#248)

* correct on simple prompt

* fix: test flashattention failed

Refactor native attention to cleanly handle 5D KV cache:
- _get_and_update_kv_cache only returns 5D fused buffer
- forward_attention receives 5D, flattens to 3D internally for attention computation
- update_kv_cache kernel accepts only 5D inputs (flattens internally for Pallas)
- Remove scattered 3D/5D dual-path compatibility code


* add bench script for temp record

* deepseekv3 accurayc aligned on num_hidden_layers=4

* update model config per model file

* revert layer

* fix redundancy

* add benchmark

* shorten

* fix

* delete comment

* add padding comment

* update

* fix lint

* update deepseek attention

* delete benchmark

* add mimo

* separate yarn of deepseek and grok

* add scripts

---------

+++ python/sgl_jax/srt/layers/attention/mla.py
@@ -77,7 +77,7 @@ class MLAAttention(nnx.Module):
                 kernel_axes=(None, None),
                 scope_name="q_a_proj",
             )
-            self.q_a_layernorm = RMSNorm(q_lora_rank, param_dtype=jnp.float32)
+            self.q_a_layernorm = RMSNorm(q_lora_rank, dtype=dtype)
             self.q_b_proj = LinearBase(
                 q_lora_rank,
                 num_heads * self.qk_head_dim,
@@ -97,7 +97,7 @@ class MLAAttention(nnx.Module):
             kernel_axes=(None, None),
             scope_name="kv_a_proj",
         )
-        self.kv_a_layernorm = RMSNorm(kv_lora_rank, param_dtype=jnp.float32)
+        self.kv_a_layernorm = RMSNorm(kv_lora_rank, dtype=dtype)
         self.kv_b_proj = LinearBase(
             kv_lora_rank,
             num_heads * (qk_nope_head_dim + v_head_dim),
